feat: 分散版をデフォルトのissue_solver_agentに変更
- issue_solver_agent_distributed.py → issue_solver_agent.py
- 実際のWorker連携システムが主力実装となったため
- 旧シミュレーション版は削除（分散版で全機能をカバー）
